+++ payments-core/src/main/java/com/stripe/android/utils/AnimationConstants.kt
@@ -1,9 +1,11 @@
-package com.stripe.android.paymentsheet.ui
+package com.stripe.android.utils
 
 import androidx.annotation.AnimRes
-import com.stripe.android.paymentsheet.R
+import androidx.annotation.RestrictTo
+import com.stripe.android.R
 
-internal object AnimationConstants {
+@RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
+object AnimationConstants {
     @AnimRes
     val FADE_IN = R.anim.stripe_paymentsheet_transition_fade_in