+++ paymentsheet/src/main/java/com/stripe/android/paymentsheet/addresselement/AddressElementActivityContract.kt
@@ -1,8 +1,10 @@
 package com.stripe.android.paymentsheet.addresselement
 
+import android.app.Activity
 import android.content.Context
 import android.content.Intent
 import androidx.activity.result.contract.ActivityResultContract
+import androidx.annotation.ColorInt
 import androidx.annotation.RestrictTo
 import androidx.core.os.bundleOf
 import com.stripe.android.core.injection.DUMMY_INJECTOR_KEY
@@ -14,9 +16,11 @@ import kotlinx.parcelize.Parcelize
 internal class AddressElementActivityContract :
     ActivityResultContract<AddressElementActivityContract.Args, AddressLauncherResult>() {
 
-    override fun createIntent(context: Context, input: Args) =
-        Intent(context, AddressElementActivity::class.java)
-            .putExtra(EXTRA_ARGS, input)
+    override fun createIntent(context: Context, input: Args): Intent {
+        val statusBarColor = (context as? Activity)?.window?.statusBarColor
+        return Intent(context, AddressElementActivity::class.java)
+            .putExtra(EXTRA_ARGS, input.copy(statusBarColor = statusBarColor))
+    }
 
     override fun parseResult(resultCode: Int, intent: Intent?) =
         intent?.getParcelableExtra<Result>(EXTRA_RESULT)?.addressOptionsResult
@@ -34,7 +38,8 @@ internal class AddressElementActivityContract :
     data class Args internal constructor(
         internal val publishableKey: String,
         internal val config: AddressLauncher.Configuration?,
-        @InjectorKey internal val injectorKey: String = DUMMY_INJECTOR_KEY
+        @InjectorKey internal val injectorKey: String = DUMMY_INJECTOR_KEY,
+        @ColorInt internal val statusBarColor: Int? = null
     ) : ActivityStarter.Args {
 
         companion object {